I have decided to keep file system discovery as a separate program. This has many advantages.
This code now references code from server. There must be a fancier
makefile which will notice changes that should remake fsmain.

The DB structure is radically changed - right down to the schema name
which is now pas2.
